Detect and avoid destructing context variables
Destructing context variables is undefined behaviour. Detect and avoid
destructing them in straightline_cleanup.

Calling destruct on a context variable does not clear it if the variable
is mentioned in the goal. Use this for detection.

Calling destruct on a context variable that is not mentioned in the goal
does sometimes clear it, and sometimes is doesn't. Because of this the
before existing code had the effect of destructing context variables,
but only very late in subproofs. This was quite slow:
fiat-crypto .vo file buildtime before: TODO
after this change: TODO